move create and validate temp timeline dir to timelines_global_map
